+++ src/themes/provider/index.ts
@@ -1,3 +1,3 @@
-export * from "./globalStyle";
+export * from "./GlobalStyle";
 export * from "./theme.provider";
 export * from "./theme.overrides";