+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Set favicon from theme provider according to overrides
+
+### Changed
+
+- Theme provider optimizations
+- Bumped `@ubloimmo/front-tokens` to `0.0.17`
+- Rewrite `GlobalStyle` in order to decrease excessive renders and increase perf
+
+### Fixed
+
+- Effect color mapping thanks to updated design tokens
+
 ## 0.6.0 - 2024-05-28
 ### Added
 - `v0.0.1` GridItem layout